+++ src/howler.core.js
@@ -562,6 +562,9 @@
           sound._seek = self.seek(ids[i]);
           sound._paused = true;
 
+          // Stop currently running fades.
+          self._stopFade(ids[i]);
+
           if (self._webAudio) {
             // make sure the sound has been created
             if (!sound._node.bufferSource) {
@@ -624,6 +627,9 @@
           sound._seek = sound._start || 0;
           sound._paused = true;
           sound._ended = true;
+
+          // Stop currently running fades.
+          self._stopFade(ids[i]);
 
           if (self._webAudio && sound._node) {
             // make sure the sound has been created
@@ -725,7 +731,7 @@
         } else {
           vol = parseFloat(args[0]);
         }
-      } else if (args.length === 2) {
+      } else if (args.length >= 2) {
         vol = parseFloat(args[0]);
         id = parseInt(args[1], 10);
       }
@@ -755,6 +761,11 @@
 
           if (sound) {
             sound._volume = vol;
+
+            // Stop currently running fades.
+            if (!args[2]) {
+              self._stopFade(id[i]);
+            }
 
             if (self._webAudio && sound._node && !sound._muted) {
               sound._node.gain.setValueAtTime(vol * Howler.volume(), ctx.currentTime);
@@ -810,7 +821,8 @@
             sound._node.gain.linearRampToValueAtTime(to, end);
 
             // Fire the event when complete.
-            setTimeout(function(id, sound) {
+            sound._timeout = setTimeout(function(id, sound) {
+              delete sound._timeout;
               setTimeout(function() {
                 sound._volume = to;
                 self._emit('faded', id);
@@ -824,7 +836,7 @@
             
             (function() {
               var vol = from;
-              var interval = setInterval(function(id) {
+              sound._interval = setInterval(function(id, sound) {
                 // Update the volume amount.
                 vol += (dir === 'in' ? 0.01 : -0.01);
 
@@ -836,17 +848,42 @@
                 vol = Math.round(vol * 100) / 100;
 
                 // Change the volume.
-                self.volume(vol, id);
+                self.volume(vol, id, true);
 
                 // When the fade is complete, stop it and fire event.
                 if (vol === to) {
-                  clearInterval(interval);
+                  clearInterval(sound._interval);
+                  delete sound._interval;
                   self._emit('faded', id);
                 }
-              }.bind(self, ids[i]), stepLen);
+              }.bind(self, ids[i], sound), stepLen);
             })();
           }
         }
+      }
+
+      return self;
+    },
+
+    /**
+     * Internal method that stops the currently playing fade when
+     * a new fade starts, volume is changed or the sound is stopped.
+     * @param  {Number} id The sound id.
+     * @return {Howl}
+     */
+    _stopFade: function(id) {
+      var self = this;
+      var sound = self._soundById(id);
+
+      if (sound._interval) {
+        clearInterval(sound._interval);
+        delete sound._interval;
+        self._emit('faded', id);
+      } else if (sound._timeout) {
+        clearTimeout(sound._timeout);
+        delete sound._timeout;
+        sound._node.gain.cancelScheduledValues(ctx.currentTime);
+        self._emit('faded', id);
       }
 
       return self;